// src/graph/types.ts
export type Input<T> = { input: T };
export type Where<T> = { where: T };

export type Maybe<T> = T | null;
// ...
export interface CreateIntent {
  agentId: number;

  name: string;

  parentId: Maybe<number>;

  triggers: CreateIntentTriggerDataInput[];

  outputs: Json[];
}

export interface CreateIntentTriggerDataInput {
  type: IntentTriggerType;

  value: string[];
}

export interface UpdateIntent {
  id: number;

  name: Maybe<string>;

  isTopLevel: Maybe<boolean>;

  parentId: Maybe<number>;

  triggers: Maybe<IntentTriggerDataInput[]>;

  outputs: Maybe<Json[]>;
}

export interface IntentTriggerDataInput {
  id: Maybe<number>;

  type: Maybe<IntentTriggerType>;

  value: Maybe<string[]>;

  actionType: ActionType;
}

export interface DeleteIntent {
  id: number;

  withChildren: boolean;
}
// ...
export enum IntentTriggerType {
  Plain = "PLAIN",
  Combination = "COMBINATION",
}

export enum ActionType {
  Create = "CREATE",
  Update = "UPDATE",
  Delete = "DELETE",
}

export type Json = any;